The exception when selecting the song was solved.
The exception of the song selection was because of the currentSongColorView, which is set at  e193bace032451881f8ae64ac1b257a99db44d83 of the feature/highlighting-the-current-song branch (#53). For the playing queue screen, there is the currentSongColorView in the layout, so the findViewById(R.id.currentSongColorView) returned a non-null value. But for the song screen, including the album detail screen, there is not the currentSongColorView in the layout, so the findViewById(R.id.currentSongColorView) returned a null value.