Revert to the original code of the VolumeFragment.kt.
While developing the precise volume, I came up with the idea to set the max value to the system volume level. Doing this, the precise volume change can be done only by changing the slider value 0.0-1.0 of the MediaPlayer. So VOLUME_MAX_VALUE_TO was not required.
Additionally, by implementing this idea, I noticed the volume warning was not required, because the max value is limited within the reasonable value.
Though, What a waste of time that was!